+++ validx/py/numbers.py
@@ -80,7 +80,7 @@ class Int(abstract.Validator):
     def __call__(self, value, __context=None):
         if value is None and self.nullable:
             return value
-        if not isinstance(value, int):
+        if not isinstance(value, int) or isinstance(value, bool):
             if isinstance(value, float) and value.is_integer():
                 # Implicitly convert ``float`` to ``int``,
                 # if the value represents integer number
@@ -183,7 +183,7 @@ class Float(abstract.Validator):
         if value is None and self.nullable:
             return value
         if not isinstance(value, float):
-            if isinstance(value, int):
+            if isinstance(value, int) and not isinstance(value, bool):
                 # Always implicitly convert ``int`` to ``float``
                 value = float(value)
             elif not self.coerce: